refactor: enable NRTs for Patient services
Enable nullable reference types for Patient related services and their
usage in controllers.

- The PatientFormView inputs should all be nullable, but the services
using it as an input expected non-nullable values. We now convert the
form to new Patient and Address DTOs with extension methods that use a
new Guard class to validate the form inputs as not null.
- Use patientDTO  in PatientService
- Update queries that might default to have nullable return values
- Update PatientInfoViewModel and PatientViewService to expect a
nullable MessagingLinkId

- Update PatientsController to handle potential null returns now in
services
- Add staffId null check for staffId user claim

+++ src/PatientPortal/Controllers/PatientsController.cs
@@ -9,6 +9,7 @@ using PatientPortal.Authorization;
 using PatientPortal.Interfaces;
 using PatientPortal.Models;
 using PatientPortal.Extensions;
+using PatientPortal.DTOs;
 
 namespace PatientPortal.Controllers
 {
@@ -32,6 +33,11 @@ namespace PatientPortal.Controllers
         [HttpGet("")]
         public IActionResult PatientManager(PatientSearch searchBar, Paginator paginationSettings)
         {
+            if(!staffId.HasValue)
+            {
+                return StatusCode(500, "Unable to retrieve staff information. Please ensure you are logged in and try again.");
+            }
+
             PatientManagerView viewModel = _patientViewService.ReturnPatientManagerView(searchBar, paginationSettings, (int)staffId);
 
             return View("PatientManager", viewModel);
@@ -48,26 +54,42 @@ namespace PatientPortal.Controllers
         {
             if (ModelState.IsValid)
             {
-                if (!_patientService.DoesPatientExist(patientFormView))
+                try
                 {
+                    var patientDTO = patientFormView.ToPatientDTO();
+                    
+                    if (!_patientService.DoesPatientExist(patientDTO))
+                    {
 
-                    int patientId = _patientService.CreatePatient(patientFormView);
+                        int patientId = _patientService.CreatePatient(patientDTO);
 
-                    return RedirectToAction("PatientInfo", "Patients", new { patientId = patientId });
+                        return RedirectToAction("PatientInfo", "Patients", new { patientId = patientId });
+                    }
+                    else
+                    {
+                        //TODO: Replace with proper ModelState error and display on form
+                        ViewBag.AlreadyExistsError = "A patient already exists with this information.";
+                    }
                 }
-
-                else
+                catch
                 {
-                    ViewBag.AlreadyExistsError = "A patient already exists with this information.";
+                    // Log the exception (not implemented here)
+                    ModelState.AddModelError(string.Empty, "An unexpected error occurred while creating the visit.");
                 }
+                
+
             }
-            return View("PatientForm");
+            return View("PatientForm", patientFormView);
         }
 
         [HttpGet("{patientId}")]
         public IActionResult PatientInfo(int patientId)
         {
-            PatientInfoViewModel patientInfo = _patientViewService.GetPatientInfo(patientId);
+            PatientInfoViewModel? patientInfo = _patientViewService.GetPatientInfo(patientId);
+            if (patientInfo == null)
+            {
+                return NotFound();
+            }
 
             return View("PatientInfo", patientInfo);
         }
@@ -84,6 +106,11 @@ namespace PatientPortal.Controllers
         [HttpPost("{patientId}/join")]
         public IActionResult MedicalTeamJoin(int patientId)
         {
+            if(!staffId.HasValue)
+            {
+                return StatusCode(500, "Unable to retrieve staff information. Please ensure you are logged in and try again.");
+            }
+            
             _patientStaffConnectionService.AddStaffToPatientTeam(patientId, (int)staffId);
             
             return RedirectToAction("PatientInfo", "Patients", new { patientId = patientId });
@@ -91,6 +118,11 @@ namespace PatientPortal.Controllers
         [HttpPost("{patientId}/leave")]
         public IActionResult MedicalTeamLeave(int patientId)
         {
+            if(!staffId.HasValue)
+            {
+                return StatusCode(500, "Unable to retrieve staff information. Please ensure you are logged in and try again.");
+            }
+            
             _patientStaffConnectionService.RemoveStaffFromPatientTeam(patientId, (int)staffId);
 
             return RedirectToAction("PatientInfo", "Patients", new { patientId = patientId });
